+++ package.json
@@ -86,11 +86,13 @@
     "commit-msg": "node ./hooks/commit-msg.js",
     "pre-commit": "lint-staged",
     "postinstall": "patch-package && nx run-many --target=postinstall --all",
-    "release": "npm run nx:graph && npm run release:phase0 && npm run release:phase1 && npm run release:phase2",
+    "release:npm": "npm run nx:graph && npm run release:phase0 && npm run release:phase1 && npm run release:phase2",
+    "release:git": "npx -p=@coveord/release git-publish-all",
+    "release:github": "npx -p=@coveord/release github-publish",
     "nx:graph": "nx graph --file=topology.json",
     "release:phase0": "npx -p=@coveord/release git-lock",
     "release:phase1": "nx run-many --target=release:phase1 --all --parallel=false --output-style=stream",
-    "release:phase2": "npx -p=@coveord/release git-publish-all"
+    "release:phase2": "npx -p=@coveord/release is-cli-release"
   },
   "workspaces": [
     "packages/*",
